+++ GUI/Elements/Button/UIIncriment.cs
@@ -4,19 +4,23 @@ using System;
 public partial class UIIncriment : VBoxContainer
 {
     // TODO: modify keys
-    public Action<float> callback;
+    public Action<double> callback;
     public override void _Ready()
     {
         GetNode("MoveUp").Connect("pressed", new Callable(this, "Incriment"));
         GetNode("MoveDown").Connect("pressed", new Callable(this, "Decriment"));
     }
-    void Incriment(){
-        if (callback != null){
+    void Incriment()
+    {
+        if (callback != null)
+        {
             callback(1f);
         }
     }
-    void Decriment(){
-        if (callback != null){
+    void Decriment()
+    {
+        if (callback != null)
+        {
             callback(-1f);
         }
     }
